Dock the probe even if the offset is too large
When the calculated offset is higher than the threshold, the script raises an error and aborts, leaving the probe attached to the toolhead. This is not a safe condition, and on printers like the V2.4, it is not easily recognized by the operator as the probe is under and behind the toolhead.

The change stows the probe before raising the error, leaving the printer in a safe condition.