fix: Use probe-manager for zero ref
With recent changes to Klipper's probe.py, the location of some
attributes has changed, in this case, the accessing of `[probe]`
`zero_reference_position`.

Before this change, setting only the `[probe]` `zero_reference_position`
and not passing `BED_POSITION` to `CALIBRATE_Z` or defining the config
option, an error would be thrown saying "Cannot find the bed position!".